Integrate disjoint intersection regressions into existing tests
